# backend/openapi.yml
openapi: 3.0.0
info:
  title: MEH
  version: 1.0.0
  description: REST API for the MEH application
servers:
  - url: /api
paths:
...
  /qrcodes:
    post:
      summary: Create a new QR code
      description: Create a new QR code for the current user
      operationId: createQrCode
      tags:
        - qrcode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QrCodeCreateRequest"
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QrCode"
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppError"
        401:
          $ref: "#/components/responses/Unauthorized"
        500:
          $ref: "#/components/responses/InternalServerError"
    get:
      summary: Get all QR codes
      description: Retrieve all QR codes belonging to the current user
      operationId: getQrCodes
      tags:
        - qrcode
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QrCode"
        401:
          $ref: "#/components/responses/Unauthorized"
        500:
          $ref: "#/components/responses/InternalServerError"
  /qrcodes/{id}:
    get:
      summary: Get a QR code by ID
      description: Retrieve a specific QR code by its unique identifier
      operationId: getQrCodeById
      tags:
        - qrcode
      parameters:
        - name: id
          in: path
          description: QR code unique identifier
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QrCode"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppError"
        500:
          $ref: "#/components/responses/InternalServerError"
  /qrcodes/by-slug/{slug}:
    get:
      summary: Get a QR code by Slug
      description: Retrieve a specific QR code by its unique slug
      operationId: getQrCodeBySlug
      tags:
        - qrcode
      parameters:
        - name: slug
          in: path
          description: QR code unique slug
          required: true
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QrCode"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppError"
        500:
          $ref: "#/components/responses/InternalServerError"
...
components:
  responses:
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AppError"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AppError"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AppError"
  schemas:
    AppError:
      type: object
      properties:
        code:
          type: string
        api_message:
          type: string
        status_code:
          type: integer
          format: int32
        values:
          type: object
          additionalProperties: true
      required:
        - code
        - status_code
...
    QrCodeCreateRequest:
      type: object
      properties:
        slug:
          type: string
          description: Custom slug for the QR code (optional)
    QrCode:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: QR code unique identifier
        owner_id:
          type: string
          format: uuid
          description: ID of the user who owns this QR code
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
        slug:
          type: string
          description: Unique slug for accessing this QR code
      required: [id, owner_id, created_at, updated_at, slug]
...